+++ engines/montecarlo.c
@@ -7,7 +7,7 @@
 #include "board.h"
 #include "engine.h"
 #include "move.h"
-#include "../joseki.h"
+#include "joseki/joseki.h"
 #include "playout/moggy.h"
 #include "playout/light.h"
 #include "engines/montecarlo.h"
