+++ joseki/Makefile
@@ -1,7 +1,10 @@
+INCLUDES=-I..
+
+OBJS := joseki.o joseki_engine.o josekiscan_engine.o
 
 SGF = joseki_44.sgf joseki_34.sgf joseki_54.sgf joseki_33.sgf 
 
-all: ../joseki19.gtp
+all: ../joseki19.gtp lib.a
 
 ../joseki19.gtp: $(SGF)
 	@echo -n "" > $@.tmp
@@ -12,4 +15,8 @@ all: ../joseki19.gtp
 	@mv $@.tmp $@
 	@echo wrote $@
 
+all: lib.a
+lib.a: $(OBJS)
 
+
+-include ../Makefile.lib
